Added new source reconstruction subpackage.
