+++ osl/README.md
@@ -5,9 +5,8 @@ A rough list so far:
 - preprocessing : MEEG data processing and cleaning (similar to OPT)
 - maxfilter : interface for Megin's commandline maxfilter
 - report : create summaries and data quality/acquisition checks from fif files
-- rhino : coregistration and source construction
-- parcellation : parcellation
-- utils : finding and organsing files, things could be used for all of the above, BIDS?
+- source_recon : coregistration, beamforming and parcellation
+- utils : finding and organsing files, things could be used for all of the above
 
 To come:
