--- osl/parcellation/__init__.py
@@ -1,3 +1,0 @@
-#!/usr/bin/python
-
-from .parcellation import *  # noqa: F401, F403
