--- osl/rhino/__init__.py
@@ -1,4 +1,0 @@
-#!/usr/bin/python
-
-from .rhino import *  # noqa: F401, F403
-from .batch import run_coreg_batch
