+++ osl/source_recon/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/python
+
+from .batch import *  # noqa: F401, F403
+from .rhino_utils import setup_fsl  # noqa: F401, F403
