+++ osl/source_recon/files/fmri_d100_parcellation_with_PCC_reduced_2mm.nii.gz
