+++ osl/source_recon/files/fmri_d100_parcellation_with_PCC_reduced_2mm_ss5mm_ds8mm_8mm.nii.gz
